+++ CHANGELOG.md
@@ -14,6 +14,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Removed
 ### Fixed
 - `weather` event did not use `duration`
+- `lever` event does not work with redstone
 ### Security
 
 ## [2.0.0] - 2023-12-31